+++ src/components/BookContainer.js
@@ -1,22 +1,22 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import BooksList from './BookList';
+import BookList from './BookList';
 import BookForm from './BookForm';
-import { getBooks } from '../redux/books/books';
+import api from '../redux/books/api';
 
 const BookContainer = () => {
   // get data from react state
   const books = useSelector((state) => state.books);
-
+  console.log(books);
   const dispatch = useDispatch();
 
   useEffect(() => {
-    dispatch(getBooks());
+    dispatch(api.fetchBooks());
   }, []);
 
   return (
     <>
-      <BooksList books={books} />
+      <BookList books={books} />
       <BookForm />
     </>
   );
