+++ src/redux/configureStore.js
@@ -1,13 +1,16 @@
-import { applyMiddleware, configureStore } from '@reduxjs/toolkit';
-import thunk from 'redux-thunk';
+import { configureStore, combineReducers } from '@reduxjs/toolkit';
 import booksReducer from './books/books';
 import categoriesReducer from './categories/categories';
 
-const store = configureStore({
-  reducer: {
+const rootReducer = combineReducers(
+  {
     books: booksReducer,
     categories: categoriesReducer,
   },
-}, applyMiddleware(thunk));
+);
+
+const store = configureStore({
+  reducer: rootReducer,
+});
 
 export default store;
